+++ arch/arc/core/cache.c
@@ -20,7 +20,7 @@
 #include <cache.h>
 #include <linker/linker-defs.h>
 #include <arch/arc/v2/aux_regs.h>
-#include <nano_internal.h>
+#include <kernel_internal.h>
 #include <misc/__assert.h>
 #include <init.h>
