+++ arch/arm/core/cortex_m/prep_c.c
@@ -20,7 +20,7 @@
 #include <zephyr/types.h>
 #include <toolchain.h>
 #include <linker/linker-defs.h>
-#include <nano_internal.h>
+#include <kernel_internal.h>
 #include <arch/arm/cortex_m/cmsis.h>
 #include <string.h>
