+++ arch/arm/include/kernel_arch_data.h
@@ -31,7 +31,7 @@ extern "C" {
 
 #ifndef _ASMLANGUAGE
 #include <kernel.h>
-#include <nano_internal.h>
+#include <kernel_internal.h>
 #include <zephyr/types.h>
 #include <misc/dlist.h>
 #include <atomic.h>
