+++ arch/nios2/include/kernel_arch_data.h
@@ -32,7 +32,7 @@ extern "C" {
 
 #ifndef _ASMLANGUAGE
 #include <kernel.h>
-#include <nano_internal.h>
+#include <kernel_internal.h>
 #include <zephyr/types.h>
 #include <misc/util.h>
 #include <misc/dlist.h>
