+++ arch/posix/soc/inf_clock/soc.c
@@ -32,7 +32,7 @@
 #include "posix_soc.h"
 #include "posix_board_if.h"
 #include "posix_core.h"
-#include "nano_internal.h"
+#include "kernel_internal.h"
 
 #define POSIX_ARCH_SOC_DEBUG_PRINTS 0
