+++ arch/riscv32/core/prep_c.c
@@ -18,7 +18,7 @@
 #include <stddef.h>
 #include <toolchain.h>
 #include <kernel_structs.h>
-#include <nano_internal.h>
+#include <kernel_internal.h>
 
 /**
  *
