+++ arch/x86/include/kernel_arch_data.h
@@ -34,7 +34,7 @@
 
 #ifndef _ASMLANGUAGE
 #include <kernel.h>
-#include <nano_internal.h>
+#include <kernel_internal.h>
 #include <zephyr/types.h>
 #include <misc/dlist.h>
 #endif
