+++ arch/xtensa/include/kernel_arch_data.h
@@ -31,7 +31,7 @@ extern "C" {
 
 #if !defined(_ASMLANGUAGE) && !defined(__ASSEMBLER__)
 #include <kernel.h>            /* public kernel API */
-#include <nano_internal.h>
+#include <kernel_internal.h>
 #include <zephyr/types.h>
 #include <misc/dlist.h>
 #include <misc/util.h>
