+++ kernel/include/kernel_internal.h
