+++ kernel/thread_abort.c
@@ -12,7 +12,7 @@
 
 #include <kernel.h>
 #include <kernel_structs.h>
-#include <nano_internal.h>
+#include <kernel_internal.h>
 #include <string.h>
 #include <toolchain.h>
 #include <linker/sections.h>
