+++ tests/kernel/mem_protect/x86_mmu_api/src/userbuffer_validate.c
@@ -11,7 +11,7 @@
 #include <linker/linker-defs.h>
 #include <arch/x86/arch.h>
 #include <ztest.h>
-#include <nano_internal.h>
+#include <kernel_internal.h>
 
 #define SKIP_SIZE 5
 #define BUFF_SIZE 10
